Start .gitignore from scratch with only what is required
The .gitignore that came original with the coding challenge was prepopulated. It looks to me like a .gitignore template for a JavaScript project from GitHub [0]. I prefer to the .gitignore just to contain exactly what is required so I deleted it all and added ignores for what I need right now. Later I will add entries as they become needed.

[0]: https://github.com/github/gitignore