Added automatic source code formatting with Prettier and Husky
Prettier is an opinionated source code formatter. It's great. It consistently formats all source code which improve readability in a subbtle way and in teams eliminates low-value code standard comments in code reviews.

Husky allows us to run stuff on git hooks. In our setup we run prettier on the pre-commit hook. That means that the repository has an enforced coding style according to the Prettier settings. You can't commit code that doesn't adhere to the standard style. Unless you add exceptions as comments or bypass the pre-commit hook.

I opted for "Option 2" in the Prettier docs on adding pre-commit hooks. I've used it before with good results.

Since I'm using WebStorm IDE I also added file watches that run Prettier on every saved file (JavaScript, TypeScript, JSX, JSON). Since I do not track the WebStorm .idea/ settings folder in the repository the changes are not seen in this repository. See the link below on how to add file watches with Prettier.

See:
https://prettier.io/
https://prettier.io/docs/en/webstorm.html#running-prettier-on-save-using-file-watcher
https://prettier.io/docs/en/precommit.html#option-2-pretty-quick-https-githubcom-azz-pretty-quick