Moved contents of App.tsx into source/Main.tsx
The plan is to have all components be located in the source/ folder. From what I can tell Expo expects the entry point of the app to be the component App located in the root of the project. Now App.tsx has basically become a proxy shell for Main.tsx.

As more files get added to the project having a separate source/ folder will make things a bit more organised I think.