Refactored TypeScript types slightly & added docs on sexuality & gender
Pulled out gender and sexuality into type aliases and then re-use those types on the user "info" and "associated" attributes. This means a tiny bit more structural consistency if the types would be made stronger at a later stage.

The gender type only takes on the values of "female" & "male" in the set of 100 users but I'm leaving it intentionally as a general string for flexibility. Same goes for the `sexuality` attribute.