Made "photos" attrivute on User type optional as per description
Noticed the following in the original challenge description provided by Feeld:

```
"photos": [ // this will be a list of zero or more photos
        {
            "url": "...",
            "width": 716,
            "height": 716
        }
    ]
```

Before this change the type asserted that the `photos` array would be non-empty. It can now be empty. Of note is that of the 100 test users all of them have a photo.