Images are now shown one after another on the user view
UserView component now displays user images one after another if the user has images. For now this is just a very rough implementation with a static 500 unit height of images. I discovered that images can range wildly in sizes and aspect ratios. I played around a bit with the different `resizeMode` props on the <Image> component. I think "cover" is the default (so it doesn't have to be specified as a prop) and it was the only one that looked remotely okay. However it's this is far from perfect. It faces appear on the boundaries of images they could be cropped out. This is something that in practice I would say should be solved on the backend. Better yet, by the user at the time of upload to give a preview of how the image will roughly turn out.

Fixed a minor spelling mistake on a variable, `interetsView` -> `interestView`.