Added a new app logo and splash screen
The new logo is a epitrochoid curve (R=3, r=1, d=0.5). It resembles a heart split into three parts but very smoothed out in a subtle way. The shape is also formed by a bean curve that cuts out the middle.  The background is a pink/purple radial gradient background. The

I haven't really tested this yet. The splash screen looks like it doesn't fit on my phone.

See:
https://en.wikipedia.org/wiki/Epitrochoid
https://en.wikipedia.org/wiki/Quartic_plane_curve