Start using Feeld API again instead of locally cache copy of users
I had a chat with Dario and briefed him that while developing I was using a cached copy of user data. He said that there was no worry about hammering the API and said that he would prefer if I'm hitting the API. So in this change I start fetching the data from the API again.