Renamed Main.tsx to ListOfUsers.tsx
I am preparing to add more views to the app. To preserve history of the list of users I'm renaming it. I will then add a new Main.tsx file which will then later on probably serve as the new entry point to the app. Most likely serving as a navigation router.