Created a new type alias UserID so semantic meaning would be clearer
In code that I am working on I wanted to create a Map of users where the key would be the user ID. Before this change it was simply a string type. By creating a type alias for it I can now refer to it in other places in the code. It's still just a string type but this way I hope the meaning and intention of the code is more clear.

That is, I find this:

`const undecidedUsers: Map<UserID, User> = new Map();`

to be more clear than this:

`const undecidedUsers: Map<string, User> = new Map();`