Added a new file with mock user objects for testing
There are 3 sample users: Ruth Bader Ginsburg, Harvey Milk and Alan Turing. These 3 people are past public figures, not real users. The main purpose of this file is to document and provide an example for the structure and content of the user object. A second reason is to provide an easy way to test instead of having to deal with API calls and promise resolution.