Published the app to Expo under my account @jonrh
Published the app to my Expo account. It should now be available at https://expo.io/@jonrh/akur

When doing the publishing it seems that Expo updated app.json (added description and a GitHub url). The assets/splash.png file also got updated. I didn't update it myself. I think this was some automated optimisation by the Expo client.

I'm not entirely sure what the `.expo-shared/assets.json`file does. I didn't find anything about it in the Expo docs. Just found an Expo blogpost that states it relates to asset optimisation:

```
The first time you run this command in an Expo project, this will create a file named assets.json in a .expo-shared directory. As the name implies, this directory is meant to be shared. You should commit this file to git so that asset state is consistent between collaborators/different machines and images are never compressed more than once.
```

It says it should be committed to Git so that's why I'm doing so.

See:
https://blog.expo.io/image-compression-with-expo-cli-d32d15cc8b73?gi=8f35061d42e1