Use demo users instead of the ones harvested from the Feeld API
When I started this project I harvested about 100 users from the Feeld API and added the JSON to a file called users.ts. I did it to increase speed of development. But after chatting with Dario he requested that I would not publish that data. So instead I created a new file, demoUsers.ts, that conforms to the User TypeScript type but the users don't contain any sensitive data. This also fixes a bug when this project was checked out on a new computer (the file existed on the computer I built the app on on but the code failed on a computer that cloned the repo).