Added getting started instructions to readme
Added two types of getting started instructions to the readme. One to start running the app quickly. On Android phones with the Expo app the QR code that was added can be scanned and this app can be run. The Expo docs do not mention that this is possible on iOS, I think it's because technically because it's agains the Terms of Service. As I recall the iOS TOS forbid apps from drastically altering behaviour. I think it may be allowed for development purposes, but not for regular users.

Also added preliminary instructions on how to get started developing and running the project locally. Added a readme shield to indicate that this project is formatted with Prettier.