Added a new file Constants.ts for constant values used in many places
This new Constants.ts file is intended to store constant values used in multiple places throughout the app.

Note: there are two constants in FeeldAPI.ts, the session token and URL. The reason why I didn't move them to the new Constants.ts file is because they are not used anywhere else in the app. Then I think it's better to have them located where they are being used.