Basic bottom tab bar navigation in place
In this change I now have a rudimentary tab bar on the bottom. The rough plan is to have 3 tabs/screens:

+ Chat: a list of mutual matches (two users approved each other) and chat with them
+ Decide: a screen to reject/skip/approve other users
+ Settings: vague idea at the moment, but idea being to provide some debug tools like resetting the matching, so it's possible to begin again.

Currently the tab bar doesn't have any text labels, just icons in the pink/purple. The icons used are SimpleLineIcons. The Chat screen has a chat bubble icon, the Decide screen has a heart icon, and the Settings screen has an circle with an exclamation mark in the middle.

See:
https://expo.github.io/vector-icons/