Refactor: moved files into new screens/ and utilities/ folders
The number of files was increasing so I felt it was time to start sorting them by functionality.

The screens/ folder is intended for React components that represents screens or views in the app. For example the screen to decide on users. The utilities/ folder should contain various non-UI related code such as API code, TypeScript types and such.