Added a new Info screen
Originally I named the screen settings, renamed it into Info. Currently it contains the name of the app at the top, "Akur", the purple/pink logo, a brief description what the app is and who it's made by, and a non-functioning button to reset the ratings. This button is intended to allow the app to start fresh after we have rated all 100 users that the Feeld devtest API provides so we can run for another session without restarting the app.