Added a new component: SwiperDots, custom purple/pink pagination dots
When a user has more than one photo react-native-swiper would display pagination dots. By default they were blue. The theme colour of the app is purple/pink and it was annoying me to no end, even though it was a very minor detail (given that I'm running very low on time for this assignment). Created a custom pagination component that takes a colour as a prop. Now the active dots have the correct colour, which makes me happy : )