UserView now displays associated user info, minor style changes
The UserView had not yet taken into account or displayed the info of an associated user. That is, the partner of a couple account. For now I just keep it super simple and display it as such:

Name1 & Name2
19y female straight, 22 male bisexual

Refactored the listing of desires and interests into their own micro components to clear up the <UserView> render function a bit. It contains the same logic just moved to a different place to hopefully improve readability a little bit.

Added few margins to improve spacing. Made "About: ...", "Desires: ...", "Interests: ..." bold.