Code re-use refactor, re-arranged user sexuality and gender in label
A user's text label used to read before: "21y male straight". Switched it around to "21y straight male". It doesn't look right in all cases, for example "43y lesbian female" should really be "43y female lesbian". But as it is currently I find it works better overall given the data from the Feeld API. This is something that could be improved.