Changed user IDs of demo users, were conflicting with some API users
Every now and then I would get a warning in the app stating that a `key` prop was duplicated. That was probably because I had manually loaded in a demo user with every API request and every now and then it would conflict with an ID from the API. This is because I was lazy when creating the demo users and didn't change the IDs. Shame on me.